requirements.txt
streamlit>=1.28.0
boto3>=1.34.0
botocore>=1.34.0
pillow>=10.0.0
requests>=2.31.0
python-dotenv>=1.0.0
